website/requirements.txt
pip==22.3
pandas==1.3.4
tensorflow==2.10.0
streamlit==1.2.0
matplotlib==3.6.1
Pillow==9.2.0
scikit-learn==1.1.2